+++ lib/mongoid/hierarchy.rb
@@ -173,6 +173,7 @@ module Mongoid
         subclass.pre_processed_defaults = pre_processed_defaults.dup
         subclass.post_processed_defaults = post_processed_defaults.dup
         subclass.scopes = scopes.dup
+        subclass.autosaved_relations = autosaved_relations.dup
 
         # We only need the _type field if inheritance is in play, but need to
         # add to the root class as well for backwards compatibility.