+++ CHANGELOG.md
@@ -27,6 +27,8 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#2851 Fixed BigDecimal demongoization of NaN values. (nkem)
+
 * \#2848 Fixed `touch` to work when usinng short timestamps. (Arthur Neves)
 
 * \#2840 Fixed end-to-end `no_timeout` option handling.