+++ CHANGELOG.md
@@ -33,6 +33,8 @@ For instructions on upgrading to newer versions, visit
 
 * \#2840 Fixed end-to-end `no_timeout` option handling.
 
+* \#2826 Dynamic fields are now properly mongoized.
+
 * \#2822 Marshal load of relations now properly reapplies extensions.
 
 ## 3.1.1