+++ CHANGELOG.md
@@ -315,11 +315,17 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#2851 Fixed BigDecimal demongoization of NaN values. (nkem)
+
 * \#2841 Calling `delete_all` or `destroy_all` on an embeds many when in the
   middle of a parent update will now properly execute the deletion.
   (Arthur Neves)
 
 * \#2835 Fixed clearing of persistence options in uniqueness validator.
+
+* \#2826 Dynamic fields are now properly mongoized.
+
+* \#2822 Marshal load of relations now properly reapplies extensions.
 
 ## 3.0.22